# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Mini Project Management System API
  version: "1.0"
  description: |
    REST API cho hệ thống Mini Project Management System (Jira-like).
    Biên soạn dựa trên SRS-Mini-Project-Management-System.md và
    System-Architecture-Mini-Project-Management-System.md.

    ## Quy ước chung
    - Base path: `/api`
    - Mọi endpoint (trừ `/auth/login`, `/auth/refresh`) yêu cầu header
      `Authorization: Bearer <accessToken>`.
    - Response thành công luôn bọc trong `ApiResponseBase`:
      `{ "success": true, "message": null, "data": <payload> }`.
    - Response lỗi luôn theo `ErrorResponse`:
      `{ "success": false, "errorCode": "...", "message": "...", "details": [] }`.
    - Danh sách luôn hỗ trợ phân trang qua query `page` (mặc định 0) và
      `size` (mặc định 20), bọc trong `PageResponseBase` (NFR-04).
    - Định dạng ngày: `yyyy-MM-dd`. Định dạng thời gian: ISO-8601.
    - Phân quyền hai lớp theo System-Architecture mục 5.2: System Role
      (ADMIN/PM/DEV) kiểm tra ở Controller; Project Role (PM/DEV/TESTER,
      theo từng Project) kiểm tra ở Service — được ghi chú trong mô tả
      từng endpoint bên dưới.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /projects/{projectId}/sprints:
    get:
      tags: [Sprint]
      summary: Xem danh sách Sprint theo Project (FR-SPR-03*)
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200': { $ref: '#/components/responses/SprintPageResponse' }
    post:
      tags: [Sprint]
      summary: Tạo Sprint (FR-SPR-01)
      description: "Yêu cầu Project Role: PM của đúng Project"
      parameters: [ { $ref: '#/components/parameters/ProjectIdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSprintRequest' }
      responses:
        '201': { $ref: '#/components/responses/SprintResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /projects/{projectId}/sprints/{id}:
    get:
      tags: [Sprint]
      summary: Xem chi tiết Sprint (FR-SPR-03*)
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200': { $ref: '#/components/responses/SprintResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Sprint]
      summary: Cập nhật Sprint (FR-SPR-01)
      description: "Yêu cầu Project Role: PM của đúng Project. Partial update — null fields được giữ nguyên."
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSprintRequest' }
      responses:
        '200': { $ref: '#/components/responses/SprintResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/sprints/{id}/status:
    patch:
      tags: [Sprint]
      summary: Chuyển trạng thái Sprint (FR-SPR-02)
      description: >
        Yêu cầu Project Role: PM của đúng Project. Ràng buộc: một Project
        chỉ được có tối đa một Sprint ở trạng thái ACTIVE tại một thời điểm —
        request bị từ chối (400) nếu vi phạm.
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSprintStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/SprintResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ───────────────────────── TASK (FR-TASK-01..05) ─────────────────────────
  /tasks:
    post:
      tags: [Task]
      summary: Tạo Task (FR-TASK-01)
      description: >
        Yêu cầu Project Role: PM của đúng Project. Trạng thái mặc định khi
        tạo là TODO. Task Key được hệ thống tự sinh theo tiền tố Project (*).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskRequest' }
      responses:
        '201': { $ref: '#/components/responses/TaskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/search:
    get:
      tags: [Task]
      summary: Tìm kiếm Task đa điều kiện (FR-TASK-04)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - name: project
          in: query
          schema: { type: integer, format: int64 }
        - name: sprint
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/TaskStatus' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/TaskPriority' }
        - name: assignee
          in: query
          schema: { type: integer, format: int64 }
        - name: keyword
          in: query
          schema: { type: string }
          description: Tìm theo Task Key hoặc Summary
      responses:
        '200': { $ref: '#/components/responses/TaskPageResponse' }
  /tasks/{id}:
    get:
      tags: [Task]
      summary: Xem chi tiết Task (FR-TASK-05*)
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      responses:
        '200': { $ref: '#/components/responses/TaskResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Task]
      summary: Cập nhật Task (FR-TASK-02)
      description: >
        Quy tắc phân quyền cập nhật:
        - Với PM của dự án: Có quyền sửa tất cả các trường và chuyển đổi trạng thái tự do (bypass workflow).
        - Với Developer (Assignee hoặc Reporter của Task): Chỉ được phép sửa: status, description, estimateHour, dueDate. Phải tuân thủ nghiêm ngặt workflow (TODO->IN_PROGRESS->TESTING->DONE, hoặc IN_PROGRESS->TODO). Các trạng thái nhảy bậc khác bị từ chối (400).
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTaskRequest' }
      responses:
        '200': { $ref: '#/components/responses/TaskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/{id}/assign:
    patch:
      tags: [Task]
      summary: Gán Task cho thành viên (FR-TASK-03)
      description: >
        Yêu cầu Project Role: PM của đúng Project. Điều kiện: assigneeId
        phải là thành viên của cùng Project với Task — nếu không, trả 400.
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignTaskRequest' }
      responses:
        '200': { $ref: '#/components/responses/TaskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ───────────────────────── COMMENT (FR-CMT-01, FR-CMT-02) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, default: 20 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Dữ liệu không hợp lệ / vi phạm ràng buộc nghiệp vụ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden:
      description: Không đủ quyền (System Role hoặc Project Role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    SprintResponse:
      description: Thông tin Sprint
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/SprintDto' }
    SprintPageResponse:
      description: Danh sách Sprint có phân trang
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/PageResponseBase'
                      - type: object
                        properties:
                          content:
                            type: array
                            items: { $ref: '#/components/schemas/SprintDto' }
    TaskResponse:
      description: Thông tin Task
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/TaskDto' }
    TaskPageResponse:
      description: Danh sách Task có phân trang
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/PageResponseBase'
                      - type: object
                        properties:
                          content:
                            type: array
                            items: { $ref: '#/components/schemas/TaskDto' }
...
  schemas:
...
    ApiResponseBase:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, nullable: true }
    PageResponseBase:
      type: object
      properties:
        totalElements: { type: integer, example: 128 }
        totalPages: { type: integer, example: 7 }
        pageNumber: { type: integer, example: 0 }
        pageSize: { type: integer, example: 20 }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        errorCode: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: "Hour phải lớn hơn 0 và không vượt quá 24" }
        details:
          type: array
          items: { type: string }
          nullable: true

    # ---- Enum ----
...
    SprintStatus: { type: string, enum: [PLANNED, ACTIVE, CLOSED] }
    TaskType: { type: string, enum: [STORY, TASK, BUG] }
    TaskPriority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    TaskStatus: { type: string, enum: [TODO, IN_PROGRESS, TESTING, DONE] }
...
    SprintDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        sprintName: { type: string }
        goal: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status: { $ref: '#/components/schemas/SprintStatus' }
    CreateSprintRequest:
      type: object
      required: [sprintName, startDate, endDate]
      properties:
        sprintName: { type: string }
        goal: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    UpdateSprintStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/SprintStatus' }

    # ---- Task ----
    TaskDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        taskKey: { type: string, example: WEB-101 }
        projectId: { type: integer, format: int64 }
        sprintId: { type: integer, format: int64, nullable: true }
        summary: { type: string }
        description: { type: string }
        taskType: { $ref: '#/components/schemas/TaskType' }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        assigneeId: { type: integer, format: int64, nullable: true }
        assigneeName: { type: string, nullable: true }
        reporterId: { type: integer, format: int64 }
        reporterName: { type: string }
        storyPoint: { type: number, nullable: true }
        estimateHour: { type: number, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateTaskRequest:
      type: object
      required: [projectId, summary, taskType, priority, reporterId]
      properties:
        projectId: { type: integer, format: int64 }
        sprintId: { type: integer, format: int64, nullable: true }
        summary: { type: string }
        description: { type: string }
        taskType: { $ref: '#/components/schemas/TaskType' }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        assigneeId: { type: integer, format: int64, nullable: true }
        reporterId: { type: integer, format: int64 }
        storyPoint: { type: number, nullable: true }
        estimateHour: { type: number, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
    UpdateTaskRequest:
      type: object
      description: Chỉ 4 trường này được phép cập nhật (FR-TASK-02)
      properties:
        status: { $ref: '#/components/schemas/TaskStatus' }
        description: { type: string }
        estimateHour: { type: number }
        dueDate: { type: string, format: date }
    AssignTaskRequest:
      type: object
      required: [assigneeId]
      properties:
        assigneeId: { type: integer, format: int64 }

    # ---- Comment ----
...